+++ great_expectations/validator/metrics_calculator.py
@@ -1,30 +1,26 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, TypedDict
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple
 
 from great_expectations.validator.computed_metric import MetricValue
 from great_expectations.validator.metric_configuration import (
     MetricConfiguration,
     MetricConfigurationID,
 )
-from great_expectations.validator.validation_graph import ValidationGraph
+from great_expectations.validator.validation_graph import (
+    MetricsCalculatorErrorResultValue,
+    ValidationGraph,
+)
 
 if TYPE_CHECKING:
     import pandas as pd
     from typing_extensions import TypeAlias
 
     from great_expectations.execution_engine import ExecutionEngine
-    from great_expectations.validator.exception_info import ExceptionInfo
 
 logger = logging.getLogger(__name__)
 logging.captureWarnings(True)
-
-
-class MetricsCalculatorErrorResultValue(TypedDict):
-    metric_configuration: MetricConfiguration
-    exception_info: ExceptionInfo
-    num_failures: int
 
 
 _MetricsDict: TypeAlias = Dict[MetricConfigurationID, MetricValue]
